% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.xts}
\alias{select.xts}
\title{Select variables by name on xts}
\usage{
select.xts(.data, ...)
}
\arguments{
\item{...}{}
}
\description{
Select variables by name on xts
}
\examples{
library(xts)
data(sample_matrix)
sample.xts <- as.xts(sample_matrix, descr='my new xts object')
sample.xts \%>\% select(Open,Close)

}
